maas_dl/requirements.txt
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.9
Werkzeug==1.0.0
connexion[swagger-ui]==2.7.0
dash==2.0.0
dash-core-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
dash-html-components==2.0.0
plotly-express==0.4.1
dash-bootstrap-components==0.13.1